Fix: share git-cas store factory
